code cleanup & modularization
moved components to different files
